+++ packages/core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @draftlab/auth
 
+## 0.10.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @draftlab/auth-router@0.3.1
+
 ## 0.10.0
 
 ### Minor Changes